+++ graphqshell/src/Shell/Components/CommandBar.hs
@@ -17,12 +17,8 @@ import Relude hiding
   ( State,
     state,
   )
-import Shell.Components.Types hiding (Event)
-import qualified Shell.Components.Types as New
-import Shell.Continuation
+import Shell.Components.Shared
 import qualified Shell.KeyMap as KeyMap
-
-data Event a = CommandSelected a deriving (Eq, Show)
 
 data State a = State
   { _stRootKeyMap :: KeyMap.KeyMap CommandBarCommand,
@@ -85,28 +81,17 @@ resetState (State rootKeyMap _) = State rootKeyMap rootKeyMap
 
 -}
 
-updateNew ::
-  New.EventChan ->
+update ::
+  EventChan ->
   State a ->
-  BrickEvent ComponentName New.Event ->
+  BrickEvent ComponentName Event ->
   EventM ComponentName (Next (State a))
-updateNew chan state@(State rootKeyMap activeKeyMap) (VtyEvent (V.EvKey (V.KChar c) [])) =
+update chan state@(State rootKeyMap activeKeyMap) (VtyEvent (V.EvKey (V.KChar c) [])) =
   case KeyMap.matchKey activeKeyMap c of
-    (Just (KeyMap.Command _ cmd)) -> emitEvent chan state (New.KeyCommand cmd)
+    (Just (KeyMap.Command _ cmd)) -> emitEvent chan state (KeyCommand cmd)
     (Just (KeyMap.Group _ keyMap)) -> continue (State rootKeyMap keyMap)
     Nothing -> continue state
-updateNew _ s _ = continue s
-
-update ::
-  State a ->
-  BrickEvent ComponentName (Event a) ->
-  EventM ComponentName (Continuation (Event a) (State a))
-update state@(State rootKeyMap activeKeyMap) (VtyEvent (V.EvKey (V.KChar c) [])) = undefined
---case KeyMap.matchKey activeKeyMap c of
---(Just (KeyMap.Command _ cmd)) -> concurrently state (pure (CommandSelected cmd))
---(Just (KeyMap.Group _ keyMap)) -> keepGoing (State rootKeyMap keyMap)
---Nothing -> keepGoing state
-update s _ = keepGoing s
+update _ s _ = continue s
 
 {-
  __     ___
