--- graphqshell/src/Shell/State.hs
@@ -1,2 +1,0 @@
--- | Utility and interface functions to deal with component state
-module Shell.State where
